+++ src/main/java/com/revature/faculty/model/Roles.java
@@ -12,7 +12,7 @@ import javax.persistence.Table;
 public class Roles {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(name="`name`")	
 	private String name;
@@ -21,17 +21,17 @@ public class Roles {
 		
 	}
 
-	public Roles(Integer id, String name) {
+	public Roles(Long id, String name) {
 		
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
