+++ src/main/java/com/revature/faculty/service/FacultyService.java
@@ -2,22 +2,21 @@ package com.revature.faculty.service;
 
 import java.util.List;
 
-
 import com.revature.faculty.dto.InsertFacultyDto;
-
 import com.revature.faculty.model.Faculty;
 
 
 public interface FacultyService {
 	List<Faculty> getFaculty();
 	
-	Faculty get(int id);
+	Faculty get(Long id);
 	
-	void save (InsertFacultyDto dto);
+	void save (InsertFacultyDto dto);	
 	
 	
 	
-	void delete(int id);
+	void delete(Long id);
 	
-	List<Faculty> getByInstitution(int id);
+	List<Faculty> getByInstitution(Long id);
+	
 }
